requirements-test.txt
asynctest==0.13.0
coverage==4.5.1
pytest==3.9.1
pytest-aiohttp==0.3.0
pytest-cov==2.7.1
pytest-freezegun==0.3.0.post1
pytest-mock==1.10.1
requests-mock==1.6.0